Fix linux build

+++ pileup.cpp
@@ -225,7 +225,7 @@ std::vector<PosData> pileup_bams(const std::vector<std::filesystem::path> &bam_f
                                  uint32_t max_coverage,
                                  uint32_t min_base_quality,
                                  uint32_t num_threads,
-                                 uint32_t min_different) {
+                                 uint16_t min_different) {
     size_t total_size = (chromosome_lengths[chromosome_id] / CHUNK_SIZE) + 1;
 
     logger()->trace("Allocating data structures...");
